+++ apps/txt/find-controller.js
@@ -87,7 +87,7 @@
   }
 
   global.InkDeskTxtFindController = Object.freeze({
-    version: '0.20.2.22',
+    version: '0.20.2.23',
     createFindController
   });
 })(globalThis);