+++ shared/ui/document-ruler-drag-controller.js
@@ -1,7 +1,7 @@
 (function (global) {
   'use strict';
 
-  const VERSION = '0.20.2.22';
+  const VERSION = '0.20.2.23';
 
   function create(options) {
     const config = options || {};